==> implementation_using_pytorch/Preprocessing/2_build_audio_vectors.ipynb <==
<h1>Build Audio Vectors</h1>

Now that the labels have been extracted, we'll use the compiled csv (df_iemocap.csv) to split the original wav files into multiple frames

! pip install librosa
! pip install tqdm

# Try for one file first
# librosa is a python package for music and audio analysis. It provides the building blocks necessary to create music information retrieval systems.
import librosa
import os
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.style as plt
from tqdm import tqdm
import pickle
import IPython.display
import librosa.display
%matplotlib inline

file_path = '/kaggle/input/iemocapfullrelease/IEMOCAP_full_release/Session1/dialog/wav/Ses01F_impro01.wav'

y, sr = librosa.load(file_path, sr=44100)
y, sr

<h1>Looping through all the files</h1>

import pandas as pd
import math

labels_df = pd.read_csv('/kaggle/working/dialog_EmoEvaluation_labels.csv')

The following cells take some time until completely executed

# sr parameter stands for "sampling rate" and it specifies the desired sampling rate of the audio data. In this case, the value 44100 indicates a common sampling rate of 44.1 kHz, which is often used for high-quality audio.
sr = 44100
audio_vectors = {}
count =0
for sess in [1]: # Using one session due to memory constraint, can replace with range(1, 6)
    wav_file_path = '/kaggle/input/iemocapfullrelease/IEMOCAP_full_release/Session{}/dialog/wav/'.format(sess)
    orig_wav_files = os.listdir(wav_file_path)
    for orig_wav_file in tqdm(orig_wav_files):
        try:
            orig_wav_vector, _sr = librosa.load(wav_file_path + orig_wav_file, sr=sr)
            orig_wav_file, file_format = orig_wav_file.split('.')
            for index, row in labels_df[labels_df['wav_file'].str.contains(orig_wav_file)].iterrows():
                start_time, end_time, truncated_wav_file_name, emotion, val, act, dom = row['start_time'], row['end_time'], row['wav_file'], row['emotion'], row['val'], row['act'], row['dom']
                start_frame = math.floor(start_time * sr)
                end_frame = math.floor(end_time * sr)
                truncated_wav_vector = orig_wav_vector[start_frame:end_frame + 1]
                audio_vectors[truncated_wav_file_name] = truncated_wav_vector
                
                count +=1
                if count ==14:
                    break
            if count== 14:
                break
        except Exception as e:
            print(e+' in {}'.format(orig_wav_file))
    with open('/kaggle/working/audio_vectors_{}.pkl'.format(sess), 'wb') as f:
        pickle.dump(audio_vectors, f)
